requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
distro==1.9.0
Django==5.1
django-behave==0.1.6
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
jiter==0.5.0
Mako==1.3.5
MarkupSafe==2.1.5
openai==1.43.1
outcome==1.3.0.post0
packaging==24.1
parse==1.20.2
parse_type==0.6.3
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pydantic_core==2.23.2
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.2
pytest-django==4.9.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0